requirements.txt
# requirements.txt
python-telegram-bot==20.7
telethon==1.32.1
